requirements.txt
flask
flask-cors
pymupdf
pdfplumber
gunicorn
groq
python-docx